+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [Unreleased][unreleased]
 
+### Fixed
+- refactor proxy dialog styles to fix #5507
+
 ### Changed
 - update `@deltachat/stdio-rpc-server` and `deltachat/jsonrpc-client` to `2.17.0`
 - update translations (7-10-2025)
+
 
 <a id="2_15_0"></a>